Исправление модели плюс реализация удаления незакрытых ордеров с возратом остатка на сток

+++ server/controllers/orderController.js
@@ -61,9 +61,45 @@ class OrderController {
     }
     async remove(req, res, next) {
         try {
-            let { id } = req.body;
-            await Order.destroy({ where: { id } });
-            return res.json(`Order with id ${id} removed successfully`);
+            const { id } = req.params;
+            // find order in db and status check
+            const order = await Order.findOne({ where: { id } });
+            // parse items from sequelize instanse to json string
+            let orderItems = JSON.stringify(order.items);
+            // reparse to object
+            orderItems = JSON.parse(orderItems);
+            // list of item ids
+            let orderItemsIds = [];
+            // check order status before removing, cause not closed order items need to return in stock
+            if (order.status === 'Closed') {
+                // if status 'Closed' remove permamently
+                await Order.destroy({ where: { id } });
+                return res.json(`Order with id ${id} removed successfully`);
+            }
+            else {
+                // mapping of orderItems to get item ids
+                orderItems.map((item) => {
+                    orderItemsIds.push(item.id)
+                })
+                // fetching products with ids from order
+                const products = await Product.findAll({ where: { id: { [Op.in]: orderItemsIds } } });
+                // convert to json from sequelize instance
+                let productsList = JSON.stringify(products)
+                // reparse list from string to object
+                productsList = JSON.parse(productsList)
+                console.log(productsList)
+                await productsList.map((item) => {
+                    let currentItemId = item.id;
+                    // find item in order with equals ids and update products qty values in database
+                    orderItems.find((elem) => {
+                        if (elem.id === currentItemId) {
+                            item.stock = item.stock + elem.qty;
+                            Product.update({ stock: item.stock }, { where: { id: item.id } })
+                            console.log(item.stock)
+                        }
+                    })
+                })
+            }
         }
         catch (e) {
             next(ApiError.badRequest(e.message));
